requirements.txt
fastapi
uvicorn
python-docx
docxtpl
openpyxl
XlsxWriter
python-pptx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
hypothesis
pytest
